Unused classes removed. PermissionsViewEvent renamed to Event
