+++ uiscanner/src/main/java/no/nordicsemi/android/common/ui/scanner/view/error/BluetoothDisabledView.kt
@@ -29,15 +29,11 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.common.ui.scanner.permissions
+package no.nordicsemi.android.common.ui.scanner.view.error
 
-import android.Manifest
+import android.bluetooth.BluetoothAdapter
 import android.content.Context
 import android.content.Intent
-import android.net.Uri
-import android.provider.Settings
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
@@ -51,67 +47,50 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.core.content.ContextCompat
-import no.nordicsemi.android.common.theme.parseBold
 import no.nordicsemi.android.common.ui.scanner.R
 import no.nordicsemi.android.common.ui.scanner.ui.AppBar
+import no.nordicsemi.android.common.ui.scanner.view.event.Event
 
 @Composable
-internal fun LocationPermissionRequiredView(
-    isDeniedForever: Boolean,
-    onEvent: (PermissionsViewEvent) -> Unit
-) {
+internal fun BluetoothDisabledView(onEvent: (Event) -> Unit) {
     Column {
-        AppBar(stringResource(id = R.string.scanner_error)) { onEvent(NavigateUp) }
+        AppBar(stringResource(id = R.string.scanner_error)) { onEvent(Event.NavigateUp) }
 
         Column(
             verticalArrangement = Arrangement.Center,
             horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier.fillMaxSize().verticalScroll(rememberScrollState()),
+            modifier = Modifier.fillMaxSize().verticalScroll(rememberScrollState())
         ) {
             Image(
-                painter = painterResource(id = R.drawable.ic_location_off),
+                painter = painterResource(id = R.drawable.ic_bluetooth_disabled),
                 contentDescription = "",
                 modifier = Modifier.padding(16.dp)
             )
 
             Text(
-                text = stringResource(id = R.string.location_permission_title),
+                text = stringResource(id = R.string.bluetooth_disabled_title),
                 color = MaterialTheme.colorScheme.secondary
             )
 
             Spacer(modifier = Modifier.height(16.dp))
 
             Text(
-                text = stringResource(id = R.string.location_permission_info).parseBold(),
+                text = stringResource(id = R.string.bluetooth_disabled_info),
                 style = MaterialTheme.typography.bodyMedium,
                 modifier = Modifier
                     .fillMaxWidth()
-                    .padding(16.dp)
+                    .padding(16.dp),
+                textAlign = TextAlign.Center
             )
 
             Spacer(modifier = Modifier.height(16.dp))
 
-            val requiredPermissions = arrayOf(
-                Manifest.permission.ACCESS_FINE_LOCATION,
-                Manifest.permission.ACCESS_COARSE_LOCATION
-            )
-
-            val launcher = rememberLauncherForActivityResult(
-                ActivityResultContracts.RequestMultiplePermissions()
-            ) { onEvent(RefreshNavigation) }
-
-            if (!isDeniedForever) {
-                Button(onClick = { launcher.launch(requiredPermissions) }) {
-                    Text(text = stringResource(id = R.string.action_grant_permission))
-                }
-            } else {
-                val context = LocalContext.current
-                Button(onClick = { openPermissionSettings(context) }) {
-                    Text(text = stringResource(id = R.string.action_settings))
-                }
+            val context = LocalContext.current
+            Button(onClick = { enableBluetooth(context) }) {
+                Text(text = stringResource(id = R.string.action_enable))
             }
 
             Spacer(modifier = Modifier.height(16.dp))
@@ -119,19 +98,12 @@ internal fun LocationPermissionRequiredView(
     }
 }
 
-private fun openPermissionSettings(context: Context) {
-    ContextCompat.startActivity(
-        context,
-        Intent(
-            Settings.ACTION_APPLICATION_DETAILS_SETTINGS,
-            Uri.fromParts("package", context.packageName, null)
-        ),
-        null
-    )
+private fun enableBluetooth(context: Context) {
+    context.startActivity(Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE))
 }
 
 @Preview
 @Composable
-private fun LocationPermissionRequiredView_Preview() {
-    LocationPermissionRequiredView(true) { }
+private fun BluetoothDisabledView_Preview() {
+    BluetoothDisabledView { }
 }
