+++ uiscanner/src/main/java/no/nordicsemi/android/common/ui/scanner/view/event/Event.kt
@@ -29,34 +29,16 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.common.ui.scanner.bonding.view
+package no.nordicsemi.android.common.ui.scanner.view.event
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import no.nordicsemi.android.common.ui.scanner.R
+import no.nordicsemi.android.common.ui.scanner.DiscoveredBluetoothDevice
 
-@Composable
-internal fun BondingSuccessView() {
-    Column(modifier = Modifier
-        .fillMaxSize()
-        .padding(16.dp)) {
-        Text(
-            text = stringResource(id = R.string.bonding_success),
-            textAlign = TextAlign.Center
-        )
-    }
-}
+internal sealed class Event {
 
-@Preview
-@Composable
-private fun BondingSuccessViewPreview() {
-    BondingSuccessView()
+    internal object RefreshNavigation : Event()
+
+    internal object NavigateUp : Event()
+
+    internal data class DeviceSelected(val device: DiscoveredBluetoothDevice) : Event()
+
 }
